requirements.txt
openai
streamlit
python-pptx
firebase_admin
google-cloud-storage
streamlit-float